Update main.c
Switched some code blocks at the beginning as I was getting munmap errors due to not initialising the arrays correctly as I was asking for the number of particles AFTER allocating the memory. Plus some other little print statements for now to see if it works in general after these changes.